requirements.txt
python-telegram-bot[ext]
slixmpp
python-dotenv